# perf(memcomparable): optimize ser/de for decimal