--- a/src/common/src/types/decimal.rs
+++ b/src/common/src/types/decimal.rs
@@ -519,34 +519,14 @@ impl Decimal {
         }
     }
 
-    /// TODO: 1. test whether the decimal in rust, any crate, has the same behavior as PG.
-    /// 2. support memcomparable encoding for dynamic decimal.
-    pub fn mantissa_scale_for_serialization(&self) -> (i128, u8) {
-        // Since the largest scale supported by `rust_decimal` is 28,
-        // and we first compare scale, we use 29 and 30 to denote +Inf and NaN.
-        match self {
-            Self::NegativeInf => (0, 29),
-            Self::Normalized(d) => {
-                // We remark that we do not dynamic numeric, i.e. the scale of all the numeric in
-                // the system is fixed. So we don't need to do any rescale, just use
-                // the `scale` of `rust_decimal`. However, it is possible that scale
-                // may overflow during calculation as `rust_decimal`'s max scale is
-                // 28.
-                (d.mantissa(), d.scale() as u8)
-            }
-            Self::PositiveInf => (0, 30),
-            Self::NaN => (0, 31),
-        }
-    }
-
     pub fn unordered_serialize(&self) -> [u8; 16] {
         // according to https://docs.rs/rust_decimal/1.18.0/src/rust_decimal/decimal.rs.html#665-684
         // the lower 15 bits is not used, so we can use first byte to distinguish nan and inf
         match self {
             Self::Normalized(d) => d.serialize(),
-            Self::NaN => [vec![1u8], vec![0u8; 15]].concat().try_into().unwrap(),
-            Self::PositiveInf => [vec![2u8], vec![0u8; 15]].concat().try_into().unwrap(),
-            Self::NegativeInf => [vec![3u8], vec![0u8; 15]].concat().try_into().unwrap(),
+            Self::NaN => [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+            Self::PositiveInf => [2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+            Self::NegativeInf => [3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
         }
     }
 
@@ -572,6 +552,28 @@ impl Decimal {
             Self::NaN => Some(Self::NaN),
             Self::PositiveInf => Some(Self::PositiveInf),
             Self::NegativeInf => Some(Self::PositiveInf),
+        }
+    }
+}
+
+impl From<Decimal> for memcomparable::Decimal {
+    fn from(d: Decimal) -> Self {
+        match d {
+            Decimal::Normalized(d) => Self::Normalized(d),
+            Decimal::PositiveInf => Self::Inf,
+            Decimal::NegativeInf => Self::NegInf,
+            Decimal::NaN => Self::NaN,
+        }
+    }
+}
+
+impl From<memcomparable::Decimal> for Decimal {
+    fn from(d: memcomparable::Decimal) -> Self {
+        match d {
+            memcomparable::Decimal::Normalized(d) => Self::Normalized(d),
+            memcomparable::Decimal::Inf => Self::PositiveInf,
+            memcomparable::Decimal::NegInf => Self::NegativeInf,
+            memcomparable::Decimal::NaN => Self::NaN,
         }
     }
 }