I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

This PR optimizes the performance of (de)serialization for Decimal values in memcomparable.
```
serialize_decimal       time:   [191.99 ns 192.69 ns 193.58 ns]
                        change: [-64.053% -63.859% -63.651%] (p = 0.00 < 0.05)
                        Performance has improved.

deserialize_decimal     time:   [30.827 ns 30.852 ns 30.885 ns]
                        change: [-77.813% -77.743% -77.677%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Other changes:
1. remove the tailing 0 of decimal encoding.
2. remove datetime functions from the memcomparable crate, because they are trivial and can be composed by ser/de for integers.

Given that the memcomparable crate has been stable for a while, I plan to publish it to crates.io after this PR.

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Refer to a related PR or issue link (optional)